ensure snapshots are written when not using types
