+++ dev/attest/src/bench/baseline.ts
@@ -1,8 +1,6 @@
 import process from "node:process"
 import { snapshot } from "arktype/internal/utils/serialize.js"
-import type { QueuedUpdate } from "../snapshot.js"
-import { createQueuedSnapshotUpdate } from "../snapshot.js"
-import { writeUpdates } from "../writeSnapshot.js"
+import { queueSnapshotUpdate } from "../snapshot.js"
 import type { BenchAssertionContext, BenchContext } from "./bench.js"
 import type {
     MarkMeasure,
@@ -10,8 +8,6 @@ import type {
     MeasureComparison
 } from "./measure/measure.js"
 import { stringifyMeasure } from "./measure/measure.js"
-
-const queuedUpdates: QueuedUpdate[] = []
 
 export const queueBaselineUpdateIfNeeded = (
     updated: Measure | MarkMeasure,
@@ -28,19 +24,12 @@ export const queueBaselineUpdateIfNeeded = (
             `Unable to update baseline for ${ctx.qualifiedName} ('lastSnapCallPosition' was unset).`
         )
     }
-    if (queuedUpdates.length === 0) {
-        process.addListener("exit", () => {
-            writeUpdates(queuedUpdates)
-        })
-    }
-    queuedUpdates.push(
-        createQueuedSnapshotUpdate({
-            position: ctx.lastSnapCallPosition,
-            serializedValue,
-            snapFunctionName: ctx.kind,
-            baselinePath: ctx.qualifiedPath
-        })
-    )
+    queueSnapshotUpdate({
+        position: ctx.lastSnapCallPosition,
+        serializedValue,
+        snapFunctionName: ctx.kind,
+        baselinePath: ctx.qualifiedPath
+    })
 }
 
 /** Pretty print comparison and set the process.exitCode to 1 if delta threshold is exceeded */
